Renamed AssemblyInfo.exe to BuildTool.exe For some reason only the gods know, AppVeyor refuses to execute applications if they are named AssemblyInfo.exe (or similar). When attempted, the following error message appears:
This version of C:\projects\sharptail\tools\AssemblyInfo.exe is not compatible with the version of Windows you're running. Check your computer's system information and then contact the software publisher.

Which is obviously an outright lie. If the **VERY SAME BINARY** is simply renamed to a different file, then application is executed without a problem, almost as if the application IS compatible with the version of Windows.